requirements.txt
folium==0.14.0
numpy==1.26.1
pandas==2.1.2
pyodbc==5.0.1
python-decouple==3.8
scikit-learn==1.3.2
